Make xss mods available standalone
This makes the XssMods module available in a seperate file. This allows
other frameworks that implement the SafeBuffer API to use it without using
ActionView.
